+++ test/IntegrationTests/LibraryVersions.g.cs
@@ -65,7 +65,7 @@ public static class LibraryVersion
         new object[] { string.Empty }
 #else
         new object[] { "2.52.0" },
-        new object[] { "2.59.0" },
+        new object[] { "2.60.0" },
 #endif
     };
     public static readonly IReadOnlyCollection<object[]> MassTransit = new List<object[]>