+++ src/OpenTelemetry.AutoInstrumentation/NativeMethods.cs
@@ -43,6 +43,42 @@ internal static class NativeMethods
         }
     }
 
+#if NET6_0_OR_GREATER
+    public static void ConfigureNativeContinuousProfiler(bool threadSamplingEnabled, uint threadSamplingInterval, bool allocationSamplingEnabled, uint maxMemorySamplesPerMinute)
+    {
+        if (IsWindows)
+        {
+            Windows.ConfigureContinuousProfiler(threadSamplingEnabled, threadSamplingInterval, allocationSamplingEnabled, maxMemorySamplesPerMinute);
+        }
+        else
+        {
+            NonWindows.ConfigureContinuousProfiler(threadSamplingEnabled, threadSamplingInterval, allocationSamplingEnabled, maxMemorySamplesPerMinute);
+        }
+    }
+
+    public static int ContinuousProfilerReadThreadSamples(int len, byte[] buf)
+    {
+        return IsWindows ? Windows.ContinuousProfilerReadThreadSamples(len, buf) : NonWindows.ContinuousProfilerReadThreadSamples(len, buf);
+    }
+
+    public static int ContinuousProfilerReadAllocationSamples(int len, byte[] buf)
+    {
+        return IsWindows ? Windows.ContinuousProfilerReadAllocationSamples(len, buf) : NonWindows.ContinuousProfilerReadAllocationSamples(len, buf);
+    }
+
+    public static void ContinuousProfilerSetNativeContext(ulong traceIdHigh, ulong traceIdLow, ulong spanId, int managedThreadId)
+    {
+        if (IsWindows)
+        {
+            Windows.ContinuousProfilerSetNativeContext(traceIdHigh, traceIdLow, spanId, managedThreadId);
+        }
+        else
+        {
+            NonWindows.ContinuousProfilerSetNativeContext(traceIdHigh, traceIdLow, spanId, managedThreadId);
+        }
+    }
+#endif
+
     // the "dll" extension is required on .NET Framework
     // and optional on .NET Core
     private static class Windows
@@ -52,6 +88,21 @@ internal static class NativeMethods
 
         [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
         public static extern void AddDerivedInstrumentations([MarshalAs(UnmanagedType.LPWStr)] string id, [In] NativeCallTargetDefinition[] methodArrays, int size);
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
+        public static extern void ConfigureContinuousProfiler(bool threadSamplingEnabled, uint threadSamplingInterval, bool allocationSamplingEnabled, uint maxMemorySamplesPerMinute);
+
+#if NET6_0_OR_GREATER
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
+        public static extern int ContinuousProfilerReadThreadSamples(int len, byte[] buf);
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
+        public static extern int ContinuousProfilerReadAllocationSamples(int len, byte[] buf);
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
+        public static extern void ContinuousProfilerSetNativeContext(ulong traceIdHigh, ulong traceIdLow, ulong spanId, int managedThreadId);
+#endif
     }
 
     // assume .NET Core if not running on Windows
@@ -62,5 +113,19 @@ internal static class NativeMethods
 
         [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
         public static extern void AddDerivedInstrumentations([MarshalAs(UnmanagedType.LPWStr)] string id, [In] NativeCallTargetDefinition[] methodArrays, int size);
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
+        public static extern void ConfigureContinuousProfiler(bool threadSamplingEnabled, uint threadSamplingInterval, bool allocationSamplingEnabled, uint maxMemorySamplesPerMinute);
+
+#if NET6_0_OR_GREATER
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
+        public static extern int ContinuousProfilerReadThreadSamples(int len, byte[] buf);
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
+        public static extern int ContinuousProfilerReadAllocationSamples(int len, byte[] buf);
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
+        public static extern void ContinuousProfilerSetNativeContext(ulong traceIdHigh, ulong traceIdLow, ulong spanId, int managedThreadId);
+#endif
     }
 }